backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
google-adk
google-genai
